Consolidated LanUdevAuto rspec tests.
